mobius-2.0: Removed WorkRunners from MobiusLoop.
The init function now runs synchronously before ‘Mobius.loop()’ returns.
Prior to 2.0 Mobius would post an "init" message to the event
WorkRunner, which would then call the init function.

Since this also means that the First model might be updated before you
get a reference to the loop, you won't be able to observe such changes
directly. An observer will still get the latest model when it’s added
though, even if no other model changes happen. In previous versions of
Mobius you could observe any such emissions by using an asynchronous
WorkRunner, but with this change the loop is always synchronous, making
observing intermediate synchronous model changes caused be init
impossible.

Furthermore, the effect handler is now posted to synchronously. If an
effect handler wants to perform asynchronous work, it has to sort that 
out by itself.

In summary this new behavior is identical to if you prior to 2.0 would
use `WorkRunner.immediate()` as your event and effect work runners.

These changes are the first steps towards removing the WorkRunner
concept from mobius-core. There are a couple of reasons that we want to
do this:

1. WorkRunners are really not the concept that MobiusLoop needs, and it
   is in fact too general for what it is used for. It would be more
   suitable for the loop to ask for something like a MessageDispatcher
   which just posts incoming messages to another thread.

2. The loop already does not care which thread the update function is
   executed on (and is guaranteed to always execute serially), so any
   re-dispatching of events could be dealt with outside of the loop.

3. Effect handlers should usually dispatch work to other threads anyway,
   so it does not make much sense for the loop to dispatch effect
   messages just to give it to the effect handler.

4. Removing threading code from mobius-core is necessary if we ever want
   to support kotlin multi-platform.