Removed WorkRunner from MobiusLoopController.
This was achieved by making it so that it is up to the connection to
dispatch models wherever you want them. MobiusAndroid has been changed
to use DispatchingMobiusController, which just wraps the view
Connectable with a WorkRunnerConnectable and posts and thus forwards the
model to the UI thread.

The code is essentially doing the same thing as before, this is just a
refactoring that allows us to remove all WorkRunner references from
MobiusLoopController.