Replaced MessageDispatcher safeguards with DisableableConsumer.
The new Disableable interface and the Consumer/Connection/Connectable implementation of it is a set of utilities that allow you to disable message processing in each respective type.

They give you the same kind of safety guarantee that MessageDispatcher.disable() had, and introducing them makes MessageDispatcher simpler and will help refactoring it in the future.